### 2019-09-05 #1709 @whedon
[PRE REVIEW]: OnlineStats.jl: A Julia package for statistics on data streams
**Submitting author:** @joshday (<a href="http://orcid.org/0000-0002-7490-6986">Josh Day</a>)
**Repository:** <a href="https://github.com/joshday/OnlineStats.jl" target ="_blank">https://github.com/joshday/OnlineStats.jl</a>
**Version:** v1.0.0
