+++ v1.0.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#API-1","page":"API","title":"API","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"","category":"page"},{"location":"api/#","page":"API","title":"API","text":"Modules = [OnlineStats, OnlineStatsBase]\nFilter = T -> typeof(T) === DataType ? !(T <: Weight) : true","category":"page"},{"location":"api/#OnlineStats.ADADELTA","page":"API","title":"OnlineStats.ADADELTA","text":"ADADELTA(ρ = .95)\n\nAn extension of ADAGRAD.\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.ADAGRAD","page":"API","title":"OnlineStats.ADAGRAD","text":"ADAGRAD()\n\nA variation of SGD with element-wise weights generated by the average of the squared gradients.\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.ADAM","page":"API","title":"OnlineStats.ADAM","text":"ADAM(β1 = .99, β2 = .999)\n\nA variant of SGD with element-wise learning rates generated by exponentially weighted first and second moments of the gradient.\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.ADAMAX","page":"API","title":"OnlineStats.ADAMAX","text":"ADAMAX(η, β1 = .9, β2 = .999)\n\nADAMAX with momentum parameters β1, β2.  ADAMAX is an extension of ADAM.\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.AutoCov","page":"API","title":"OnlineStats.AutoCov","text":"AutoCov(b, T = Float64; weight=EqualWeight())\n\nCalculate the auto-covariance/correlation for lags 0 to b for a data stream of type T.\n\nExample\n\ny = cumsum(randn(100))\no = AutoCov(5)\nfit!(o, y)\nautocov(o)\nautocor(o)\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.BiasVec","page":"API","title":"OnlineStats.BiasVec","text":"BiasVec(x)\n\nLightweight wrapper of a vector which adds a \"bias\" term at the end.\n\nExample\n\nBiasVec(rand(5))\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.Bootstrap","page":"API","title":"OnlineStats.Bootstrap","text":"Bootstrap(o::OnlineStat, nreps = 100, d = [0, 2])\n\nCalculate an online statistical bootstrap of nreps replicates of o.  For each call to fit!, any given replicate will be updated rand(d) times (default is double or nothing).\n\nExample\n\no = Bootstrap(Variance())\nfit!(o, randn(1000))\nconfint(o, .95)\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.CallFun","page":"API","title":"OnlineStats.CallFun","text":"CallFun(o::OnlineStat, f::Function)\n\nCall f(o) every time the OnlineStat o gets updated.\n\nExample\n\no = CallFun(Mean(), println)\nfit!(o, [0,0,1,1])\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.Diff","page":"API","title":"OnlineStats.Diff","text":"Diff(T::Type = Float64)\n\nTrack the difference and the last value.\n\nExample\n\no = Diff()\nfit!(o, [1.0, 2.0])\nlast(o)\ndiff(o)\n\n\n\n\n\n","category":"type"},{"location":"api/#OnlineStats.FastForest","page":"API","title":"OnlineStats.FastForest","text"
[…]
