### 2019-09-10 328aa00
Update partition.jl (#157)
