### 2019-09-10 c27ad10
Fix docstring
