### 2019-09-27 #1709 @vchuravy
I have time to review this in about a week.
