### 2019-10-11 665cdf3
Fixed examples to explicitly iterate over observation rows
