### 2019-10-13 cade71e
Fixed examples to explicitly iterate over observation rows (#162)
