+++ src/viz/heatmap.jl
@@ -1,13 +1,13 @@
 """
     Heatmap(xedges, yedges; left = true, closed = true)
 
-Create a two dimensional histogram with the bin partition created by `xedges` and `yedges`.  
+Create a two dimensional histogram with the bin partition created by `xedges` and `yedges`.
 When fitting a new observation, the first value will be associated with X, the second with Y.
 
 - If `left`, the bins will be left-closed.
 - If `closed`, the bins on the ends will be closed.  See [Hist](@ref).
 
-# Example 
+# Example
 
     o = fit!(HeatMap(-5:.1:5, -5:.1:5), eachrow(randn(10^5, 2)))
 
@@ -15,11 +15,11 @@ When fitting a new observation, the first value will be associated with X, the s
     plot(o)
 """
 mutable struct HeatMap{EX, EY} <: OnlineStat{XY}
-    xedges::EX 
+    xedges::EX
     yedges::EY
     counts::Matrix{Int}
     out::Int
-    left::Bool 
+    left::Bool
     closed::Bool
 
     function HeatMap(x::T, y::S; left::Bool=true, closed::Bool=true) where {T<:AbstractVector,S<:AbstractVector}
@@ -47,13 +47,13 @@ function _merge!(a::HeatMap, b::HeatMap)
     if (a.xedges == b.xedges) && (a.yedges == b.yedges) && (a.left == b.left) && (a.closed == b.closed)
         a.out += b.out
         a.counts .+= b.counts
-    else 
+    else
         @warn "HeatMaps differ in edges or direction of bins being closed.  No merging occurred."
     end
 end
 
 @recipe function f(o::HeatMap)
-    seriestype --> :heatmap 
+    seriestype --> :heatmap
     z = Float64.(o.counts)
     z[z .== 0] .= NaN
     midpoints(o.xedges), midpoints(o.yedges), z
