### 2019-10-14 #1709 @ahwillia
I'm unavailable until after Nov 1, but could review then.
