### 2019-10-22 f6725f7
KMeans fix and histogram fix
