### 2019-10-22 e133543
uncomment tests
