+++ test/runtests.jl
@@ -28,165 +28,165 @@ mergevals(o1::OnlineStat, y1, y2) = map(value, mergestats(o1, y1, y2))
 
 
 
-# #-----------------------------------------------------------------------# AutoCov
-# @testset "AutoCov" begin
-#     o = fit!(AutoCov(10), y)
-#     @test autocov(o) ≈ autocov(y, 0:10)
-#     @test autocor(o) ≈ autocor(y, 0:10)
-#     @test nobs(o) == n
-# end
-# #-----------------------------------------------------------------------# Bootstrap
-# @testset "Bootstrap" begin
-#     o = fit!(Bootstrap(Mean(), 100, [1]), y)
-#     @test all(value.(o.replicates) .== value(o.stat))
-#     c = confint(o)
-#     @test length(c) == 2
-#     @test c[1] ≤ c[2]
-# end
-# #-----------------------------------------------------------------------# CallFun
-# @testset "CallFun" begin
-#     i = 0
-#     o = fit!(CallFun(Mean(), x -> i+=1), y)
-#     @test value(o) ≈ mean(y)
-#     @test i == n
-#     @test ≈(mergevals(CallFun(Mean(), x->nothing), y, y2)...)
-# end
-# #-----------------------------------------------------------------------# CovMatrix
-# @testset "CovMatrix" begin
-#     o = fit!(CovMatrix(), eachrow(ymat))
-#     @test value(o) ≈ cov(ymat)
-#     @test cov(o) ≈ cov(ymat)
-#     @test cor(o) ≈ cor(ymat)
-#     @test all(x -> ≈(x...), zip(var(o), var(ymat; dims=1)))
-#     @test all(x -> ≈(x...), zip(std(o), std(ymat; dims=1)))
-#     @test all(x -> ≈(x...), zip(mean(o), mean(ymat; dims=1)))
+#-----------------------------------------------------------------------# AutoCov
+@testset "AutoCov" begin
+    o = fit!(AutoCov(10), y)
+    @test autocov(o) ≈ autocov(y, 0:10)
+    @test autocor(o) ≈ autocor(y, 0:10)
+    @test nobs(o) == n
+end
+#-----------------------------------------------------------------------# Bootstrap
+@testset "Bootstrap" begin
+    o = fit!(Bootstrap(Mean(), 100, [1]), y)
+    @test all(value.(o.replicates) .== value(o.stat))
+    c = confint(o)
+    @test length(c) == 2
+    @test c[1] ≤ c[2]
+end
+#-----------------------------------------------------------------------# CallFun
+@testset "CallFun" begin
+    i = 0
+    o = fit!(CallFun(Mean(), x -> i+=1), y)
+    @test value(o) ≈ mean(y)
+    @test i == n
+    @test ≈(mergevals(CallFun(Mean(), x->nothing), y, y2)...)
+end
+#-----------------------------------------------------------------------# CovMatrix
+@testset "CovMatrix" begin
+    o = fit!(CovMatrix(), eachrow(ymat))
+    @test value(o) ≈ cov(ymat)
+    @test cov(o) ≈ cov(ymat)
+    @test cor(o) ≈ cor(ymat)
+    @test all(x -> ≈(x...), zip(var(o), var(ymat; dims=1)))
+    @test all(x -> ≈(x...), zip(std(o), std(ymat; dims=1)))
+    @test all(x -> ≈(x...), zip(mean(o), mean(ymat; dims=1)))
 
[…]
