### 2019-10-28 cb17972
remove .md on license
