+++ README.md
@@ -49,4 +49,4 @@ See also the list of [contributors](https://github.com/joshday/OnlineStats.jl/co
 
 ## License
 
-OnlineStats is licensed under the MIT License - see the [LICENSE.md](LICENSE.md) file for details.
+OnlineStats is licensed under the MIT License - see the [LICENSE](LICENSE) file for details.
