### 2019-11-06 #1816 @ahwillia
This looks good to me -- I went ahead and checked all my boxes. The documentation is really nice, and the functionality of the package is clear.

My only feedback is that it would be nice to show more full-fledged examples or tutorials, since the operations performed by this particular package are nice building blocks for more complex models/applications. Are there any examples of other repos using `OnlineStats.jl` to do impressive things? Can these be mentioned in the documentation or the README?

Other than this suggestion, I am happy to see this accepted and published as is.
