### 2019-11-06 #1816 @ahwillia
Also one area the documentation could use more work is the `StatLearn` object (e.g. for online fitting of GLMs). An example showing how to take multiple passes over the same dataset (i.e. multiple epochs) could be useful.
