### 2019-12-04 b65d5c2
remove background color for hero banner in docs
