+++ docs/src/index.md
@@ -1,6 +1,6 @@
 ```@raw html
 <div style="width:100%; height:150px;border-width:4px;border-style:solid;
-        border-color:#2269D1;border-radius:10px;background-color:#6BFFE4;text-align:center;">
+        border-color:#2269D1;border-radius:10px;text-align:center;">
     <h1>Star us on GitHub!</h1>
     <a class="github-button" href="https://github.com/joshday/OnlineStats.jl" data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star joshday/OnlineStats.jl on GitHub" style="margin:auto">Star</a>
     <script async defer src="https://buttons.github.io/buttons.js"></script>
