### 2020-01-02 0898fa8
algorithm docs
