+++ src/algorithms.jl
@@ -12,6 +12,8 @@ abstract type SGAlgorithm <: Algorithm end
     SGD()
 
 Stochastic Gradient Descent.
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
 """
 struct SGD <: SGAlgorithm end
 
@@ -39,6 +41,10 @@ struct SGD <: SGAlgorithm end
     ADAGRAD()
 
 A variation of [`SGD`](@ref) with element-wise weights generated by the average of the squared gradients.
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
+- Note: ADAGRAD uses a learning rate in OnlineStats, which differs from how it is typically presented.
+  See https://www.seqstat.com/post/adagrad/ for details.
 """
 mutable struct ADAGRAD <: SGAlgorithm 
     h::Vector{Float64}
@@ -60,6 +66,8 @@ end
 
 A Variation of [`ADAGRAD`](@ref) that uses element-wise weights generated by an exponentially 
 weighted mean of the squared gradients.
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
 """
 mutable struct RMSPROP <: SGAlgorithm
     α::Float64
@@ -82,6 +90,8 @@ end
     ADADELTA(ρ = .95)
 
 An extension of [`ADAGRAD`](@ref).
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
 """
 mutable struct ADADELTA <: SGAlgorithm 
     v::Vector{Float64}
@@ -112,6 +122,8 @@ end
 
 A variant of [`SGD`](@ref) with element-wise learning rates generated by exponentially weighted
 first and second moments of the gradient.
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
 """
 mutable struct ADAM <: SGAlgorithm 
     m::Vector{Float64}
@@ -141,6 +153,8 @@ end
     ADAMAX(η, β1 = .9, β2 = .999)
 
 ADAMAX with momentum parameters `β1`, `β2`.  ADAMAX is an extension of [`ADAM`](@ref).
+
+- Reference: https://ruder.io/optimizing-gradient-descent/
 """
 mutable struct ADAMAX <: SGAlgorithm
     β1::Float64
@@ -200,6 +214,8 @@ end
     MSPI()
 
 Majorized Stochastic Proximal Iteration.
+
+- Reference: https://repository.lib.ncsu.edu/handle/1840.20/34945
 """
 struct MSPI <: Algorithm end
 
@@ -207,6 +223,8 @@ struct MSPI <: Algorithm end
     OMAP()
 
 Online MM via Averaged Parameter.
+
+- Reference: https://repository.lib.ncsu.edu/handle/1840.20/34945
 """
 struct OMAP <: Algorithm end
 
@@ -214,6 +232,8 @@ struct OMAP <: Algorithm end
     OMAS()
 
 Online MM via Averaged Surrogate.
+
+- Reference: https://repository.lib.ncsu.edu/handle/1840.20/34945
 """
 mutable struct OMAS <: Algorithm  
     a::Vector{Float64}
