+++ docs/src/stats_and_models.md
@@ -1,29 +1,46 @@
 # Statistics and Models
 
+## Univariate Statistics
+
 | Statistic/Model                    | OnlineStat                 |
 |:-----------------------------------|:---------------------------|
-| **Univariate Statistics:**         |                            |
 | Mean                               | [`Mean`](@ref)             |
 | Variance                           | [`Variance`](@ref)         |
 | Quantiles                          | [`Quantile`](@ref) and [`P2Quantile`](@ref) |
 | Maximum/Minimum                    | [`Extrema`](@ref)          |
 | Skewness and kurtosis              | [`Moments`](@ref)          |
 | Sum                                | [`Sum`](@ref)              |
-| **Time Series:**                   |                            |
+
+## Time Series
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | Difference                         | [`Diff`](@ref)             |
 | Lag                                | [`Lag`](@ref)              |
 | Autocorrelation/autocovariance     | [`AutoCov`](@ref)          |
-| Tracked history                    | [`StatLag`](@ref)      |
-| **Multivariate Analysis:**         |                            |
+| Tracked history                    | [`StatLag`](@ref)          |
+
+## Multivariate Analysis
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | Covariance/correlation matrix      | [`CovMatrix`](@ref)        |
 | Principal components analysis      | [`CovMatrix`](@ref)        |
 | K-means clustering (SGD)           | [`KMeans`](@ref)           |
 | Multiple univariate statistics     | [`Group`](@ref) |
-| **Nonparametric Density Estimation:**|                          |
+
+## Nonparametric Density Estimation
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | Histograms/continuous density      | [`Hist`](@ref) and [`KHist`](@ref) |
 | Approximate order statistics       | [`OrderStats`](@ref)       |
 | Count for each unique value        | [`CountMap`](@ref)         |
-| **Parametric Density Estimation:** |                            |
+
+## Parametric Density Estimation
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | Beta                               | [`FitBeta`](@ref)          |
 | Cauchy                             | [`FitCauchy`](@ref)        |
 | Gamma                              | [`FitGamma`](@ref)         |
@@ -31,7 +48,11 @@
 | Normal                             | [`FitNormal`](@ref)        |
 | Multinomial                        | [`FitMultinomial`](@ref)   |
 | MvNormal                           | [`FitMvNormal`](@ref)      |
-| **Statistical Learning:**          |                            |
+
+## Statistical Learning
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | GLMs with regularization           | [`StatLearn`](@ref)        |
 | Logistic regression                | [`StatLearn`](@ref)        |
 | Linear SVMs                        | [`StatLearn`](@ref)        |
@@ -43,12 +64,20 @@
 | Decision Trees                     | [`FastTree`](@ref)         |
 | Random Forest                      | [`FastForest`](@ref)       |
 | Naive Bayes Classifier             | [`NBClassifier`](@ref)     |
-| **Other:**                         |                            |
+
+## Other
+
+| Statistic/Model                    | OnlineStat                 |
+|:-----------------------------------|:---------------------------|
 | Statistical Bootstrap              | [`Bootstrap`](@ref)        |
[…]
