### 2020-01-03 3427569
CompatHelper: add new compat entry for "OrderedCollections" at version "1.1"
