### 2020-01-03 2161408
CompatHelper: add new compat entry for "StatsBase" at version "0.32"
