### 2020-01-03 e72e8fc
CompatHelper: add new compat entry for "RecipesBase" at version "0.7"
