### 2020-01-03 d37c6c7
CompatHelper: add new compat entry for "SweepOperator" at version "0.3"
