### 2020-01-03 1b3d9b2
CompatHelper: add new compat entry for "Reexport" at version "0.2"
