### 2020-01-03 e1c30f1
CompatHelper: add new compat entry for "PenaltyFunctions" at version "0.1"
