### 2020-01-03 282d6cb
CompatHelper: add new compat entry for "LearnBase" at version "0.2"
