### 2020-01-03 0d7ac6b
CompatHelper: add new compat entry for "LossFunctions" at version "0.5" (#167)
