### 2020-01-03 db6b723
CompatHelper: add new compat entry for "OrderedCollections" at version "1.1" (#168)
