### 2020-01-03 6c1e4c1
CompatHelper: add new compat entry for "PenaltyFunctions" at version "0.1" (#173)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
