### 2020-01-03 c1edef5
CompatHelper: add new compat entry for "StatsBase" at version "0.32" (#169)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
