### 2020-01-03 f20c6c4
CompatHelper: add new compat entry for "RecipesBase" at version "0.7" (#170)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
