### 2020-01-03 7850bd8
CompatHelper: add new compat entry for "SweepOperator" at version "0.3" (#171)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
