### 2020-01-03 f27840e
CompatHelper: add new compat entry for "Reexport" at version "0.2" (#172)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
