+++ Project.toml
@@ -20,6 +20,7 @@ SweepOperator = "7522ee7d-7047-56d0-94d9-4bc626e7058d"
 [compat]
 LossFunctions = "0.5"
 OnlineStatsBase = "1.0"
+Reexport = "0.2"
 SweepOperator = "0.3"
 RecipesBase = "0.7"
 StatsBase = "0.32"
