### 2020-01-03 164a5c2
CompatHelper: add new compat entry for "LearnBase" at version "0.2" (#174)
Co-authored-by: Josh Day <emailjoshday@gmail.com>
