+++ Project.toml
@@ -18,6 +18,7 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 SweepOperator = "7522ee7d-7047-56d0-94d9-4bc626e7058d"
 
 [compat]
+LearnBase = "0.2"
 LossFunctions = "0.5"
 OnlineStatsBase = "1.0"
 Reexport = "0.2"
