### 2020-01-12 636b707
try to remove sitename in sidebar of docs
