+++ docs/make.jl
@@ -1,7 +1,7 @@
 using Documenter, OnlineStats, OnlineStatsBase
 
 makedocs(
-    format = Documenter.HTML(assets = ["assets/style.css"]),
+    format = Documenter.HTML(assets = ["assets/style.css"], sidebar_sitename=false),
     sitename = "OnlineStats Docs",
     modules = [OnlineStats, OnlineStatsBase],
     authors = "Josh Day",
