### 2020-01-28 #1816 @pkofod
@karthik I can review but I do have to disclose that we are both employed by Julia Computing. Though, we do live 7000km from each other, and have never worked on projects together. I have never used OnlineStats.jl either, although I'm aware of it's existence.
