### 2020-01-28 #1816 @karthik
@pkofod I have updated the checklist above with your name. Please work through that list. Reviewer guidelines are here: http://joss.theoj.org/about#reviewer_guidelines
