### 2020-01-28 #1816 @pkofod
I have looked at the repo for the first part and everything looks fine. It's structured like a typical Julia package, and has docs, CI, a proper readme with examples and a clear indication that contributions are welcome but Josh has strong preference for the design choices. I will try to find time to read the actual paper tomorrow or the day after.
