### 2020-01-31 #1816 @pkofod
Everything looks pretty good. There is one thing missing IMO, but it might be a matter of focus and/or intent. Either OnlineStats.jl is mainly seen as a platform for developing online algorithms, and in that case the example is fine: it shows how to implement an algorithm using the structure provided by the package. However, if the goal (as I understood it at least) is mainly to provide implementations, then I think a simple example of using the *existing* package functionality would be nice. This could/would also show the Plots.jl integration. Just something like the following from the docs would suffice:
```
o = Partition(Mean())
o2 = Partition(Extrema())

s = Series(o, o2)

fit!(s, y)

plot(s, layout = 1, xlab = "Nobs")
```
Other than that, lgtm
