### 2020-02-04 #1816 @karthik
Thanks @pkofod
@joshday: Can you respond to the comment above:

> Either OnlineStats.jl is mainly seen as a platform for developing online algorithms, and in that case the example is fine: it shows how to implement an algorithm using the structure provided by the package. However, if the goal (as I understood it at least) is mainly to provide implementations, then I think a simple example of using the existing package functionality would be nice.
