### 2020-02-06 #1816 @joshday
In response @ahwillia suggestion, I have added a section to the README about packages that use OnlineStats.

In response to @pkofod:

> Either OnlineStats.jl is mainly seen as a platform for developing online algorithms, and in that case the example is fine: it shows how to implement an algorithm using the structure provided by the package. However, if the goal (as I understood it at least) is mainly to provide implementations, then I think a simple example of using the existing package functionality would be nice.

The paper is meant for a more technical audience (the "developing online algorithms" crowd) whereas the documentation is meant for general users.  I'm happy to add a higher level example, but I think most people will check the documentation first.
