### 2020-02-07 b7871b3
CompatHelper: bump compat for "RecipesBase" to "0.8"
