+++ Project.toml
@@ -21,12 +21,12 @@ SweepOperator = "7522ee7d-7047-56d0-94d9-4bc626e7058d"
 LearnBase = "0.2"
 LossFunctions = "0.5"
 OnlineStatsBase = "1.0"
-Reexport = "0.2"
-SweepOperator = "0.3"
-RecipesBase = "0.7"
-StatsBase = "0.32"
-PenaltyFunctions = "0.1"
 OrderedCollections = "1.1"
+PenaltyFunctions = "0.1"
+RecipesBase = "0.7, 0.8"
+Reexport = "0.2"
+StatsBase = "0.32"
+SweepOperator = "0.3"
 julia = "1.0"
 
 [extras]
