### 2020-02-07 a4cbeef
fix paper references
