+++ paper/paper.bib
@@ -5,41 +5,21 @@
   volume={59},
   number={1},
   pages={65--98},
-  year={2017}
+  year={2017},
+  doi={https://doi.org/10.1137/141000671}
 }
 
-@BOOK{knuth_aocp2,
-    AUTHOR      = "Donald Knuth",
-    TITLE       = "The Art of Computer Programming, Volume 2: Seminumerical Algorithms",
-    EDITION     = "Third",
-    YEAR        = {1998},
-    PUBLISHER   = "Addison-Wesley"
-}
-
-@article{Ben-Haim,
- author = {Ben-Haim, Yael and Tom-Tov, Elad},
- title = {A Streaming Parallel Decision Tree Algorithm},
- journal = {J. Mach. Learn. Res.},
- issue_date = {3/1/2010},
- volume = {11},
- month = mar,
- year = {2010},
- issn = {1532-4435},
- pages = {849--872},
- numpages = {24},
- url = {http://dl.acm.org/citation.cfm?id=1756006.1756034},
- acmid = {1756034},
- publisher = {JMLR.org},
-}
-
-@ARTICLE{hunter+lange_2004,
-    AUTHOR      = "David R. Hunter and Kenneth Lange",
-    YEAR        = {2004},
-    TITLE       = "A Tutorial on MM Algorithms",
-    PAGES       = {30-37},
-    JOURNAL     = {The American Statistician},
-    VOLUME      = {58},
-    NUMBER      = {1}
+@article{Julia,
+  author = {Bezanson, Jeff. and Edelman, Alan. and Karpinski, Stefan. and Shah, Viral B.},
+  title = {Julia: A Fresh Approach to Numerical Computing},
+  journal = {SIAM Review},
+  volume = {59},
+  number = {1},
+  pages = {65-98},
+  year = {2017},
+  doi = {10.1137/141000671},
+  URL = { https://doi.org/10.1137/141000671},
+  eprint = {https://doi.org/10.1137/141000671}
 }
 
 @software{Kafka,
@@ -48,10 +28,9 @@
   url = {https://kafka.apache.org},
 }
 
-
 @software{SparkStreaming,
   author = {{Apache Software Foundation}},
-  title = {Kafka},
+  title = {Spark Streaming},
   url = {https://spark.apache.org/streaming/},
 }
 
