+++ paper/paper.bib
@@ -1,15 +1,4 @@
 @article{Julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM Review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  doi={https://doi.org/10.1137/141000671}
-}
-
-@article{Julia,
   author = {Bezanson, Jeff. and Edelman, Alan. and Karpinski, Stefan. and Shah, Viral B.},
   title = {Julia: A Fresh Approach to Numerical Computing},
   journal = {SIAM Review},
