### 2020-02-07 #1816 @joshday
I've fixed the invalid DOI.

The "missing" DOI I believe is in error, as I'm citing the actual software.  The above DOI is for a book chapter in "Beginning Apache Spark 2", which is not what I'm citing.
