### 2020-02-07 #1816 @karthik
If the software doesn't have a DOI, I think we can skip it but I'll tag @arfon in case he can help.
