### 2020-02-07 #1816 @arfon
> If the software doesn't have a DOI, I think we can skip it but I'll tag @arfon in case he can help.

This sounds fine. The DOI search/recommender is just using the Crossref API looking for similar titles.
