### 2020-02-08 #1816 @kthyng
@joshday Hi! I'll take over from here.

What version number do you want associated with this publication?

Also, can you create an archive of your code at, for example, Zenodo, and report the doi back here? Please be sure that the title and author list match your paper exactly, which might requires editing the metadata.
