### 2020-02-10 #1816 @joshday
@kthyng The version I'd like to use is is v1.0.3 and the DOI is `10.5281/zenodo.3659245`.  Let me know if there's anything else I can do to help!

A big thank you to @ahwillia and @pkofod for reviewing and @karthik for keeping this thing moving along!
