### 2020-02-11 #1816 @joshday
My apologies for missing this in the proof, but a unicode character (μ) fails to render in the example at the end of the paper.  Is it possible to revise the paper?  I can simply change the problem character to ASCII.
