### 2023-06-21 #1816 @kthyng
Oh editorialbot maybe doesn't even work because this is so old!

@openjournals/dev I think we need a manual intervention for this one?
